Addressed styling issues mentioned by @marmbrus
