+++ sql/catalyst/src/main/scala/org/apache/spark/sql/catalyst/analysis/Catalog.scala
@@ -21,8 +21,7 @@ package analysis
 
 import scala.collection.mutable
 
-import org.apache.spark.sql.catalyst.plans.logical.{Subquery, LogicalPlan}
-
+import org.apache.spark.sql.catalyst.plans.logical.{LogicalPlan, Subquery}
 
 /**
  * An interface for looking up relations by name.  Used by an [[Analyzer]].
