+++ sql/catalyst/src/main/scala/org/apache/spark/sql/catalyst/errors/package.scala
@@ -25,8 +25,9 @@ import org.apache.spark.sql.catalyst.trees.TreeNode
  */
 package object errors {
 
-  class TreeNodeException[TreeType <: TreeNode[_]]
-    (tree: TreeType, msg: String, cause: Throwable) extends Exception(msg, cause) {
+  class TreeNodeException[TreeType <: TreeNode[_]](
+      tree: TreeType, msg: String, cause: Throwable)
+    extends Exception(msg, cause) {
 
     // Yes, this is the same as a default parameter, but... those don't seem to work with SBT
     // external project dependencies for some reason.
